BuyandBye_project/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.5
django-allauth==0.44.0
django-crispy-forms==1.10.0
django-etc==1.3.0
django-extensions==3.1.0
django-hitcount==1.3.3
django-money==1.3
django-phonenumber-field==5.0.0
filelock==3.0.12
idna==3.1
isort==5.7.0
joblib==1.0.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.19.5
oauthlib==3.1.0
pandas==1.2.0
phonenumbers==8.12.16
Pillow==8.1.0
pipenv==2020.11.15
pkg-resources==0.0.0
py-moneyed==1.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.2
virtualenv==20.3.1
virtualenv-clone==0.5.4
wrapt==1.12.1
